// OMKARAUTOAPP/src/app/models/reportmodel.ts
export class Reportmodel {
    pageNumber: number = 0;
    pageSize : number = 0;
    sortColumn: string = "";
    sortOrder: string = ""; 
    search: string = "";
    fromDate: string = "";
    toDate: string = "";
    filterStr : string = "";
    filterStr1: string = "";
    filterStr2: string = "";
    filterStr3: string = "";
}